- move `dump_stats` to lib
